+++ source/Game/ball.h
@@ -1,11 +1,11 @@
 
 // struct Ball ;
 struct Ball {
-	double x;
-	double y;
+	float x;
+	float y;
 	int radius;
-	double speed;
-	double angle;
+	float speed;
+	float angle;
 };
 void draw_ball(struct Ball *self);
 void move_ball(struct Ball *self, int block_layout[][2]);
